@@ -4,7 +4,7 @@
     "license": "GPLv2",
     "type": "project",
     "require": {
-        "php": "^7.4.0",
+        "php": "^8.0.2",
         "ext-ctype": "*",
         "ext-gd": "*",
         "ext-iconv": "*",
@@ -46,7 +46,7 @@
         "symfony/console": "5.4.*",
         "symfony/dotenv": "5.4.*",
         "symfony/expression-language": "5.4.*",
-        "symfony/flex": "^1.3",
+        "symfony/flex": "^2.2",
         "symfony/form": "5.4.*",
         "symfony/framework-bundle": "5.4.*",
         "symfony/http-client": "5.4.*",
@@ -75,15 +75,15 @@
         "vich/uploader-bundle": "^1.16"
     },
     "require-dev": {
-        "codeception/codeception": "^4.1",
-        "codeception/module-asserts": "^2.0",
-        "codeception/module-datafactory": "^2.0",
-        "codeception/module-db": "^2.0",
-        "codeception/module-doctrine2": "^2.0",
-        "codeception/module-filesystem": "^2.0",
-        "codeception/module-phpbrowser": "^2.0",
-        "codeception/module-rest": "^2.0",
-        "codeception/module-symfony": "^2.1",
+        "codeception/codeception": "^5.0",
+        "codeception/module-asserts": "^3.0",
+        "codeception/module-datafactory": "^3.0",
+        "codeception/module-db": "^3.0",
+        "codeception/module-doctrine2": "^3.0",
+        "codeception/module-filesystem": "^3.0",
+        "codeception/module-phpbrowser": "^3.0",
+        "codeception/module-rest": "^3.3",
+        "codeception/module-symfony": "^3.1",
         "dealerdirect/phpcodesniffer-composer-installer": "^0.7",
         "doctrine/doctrine-fixtures-bundle": "^3.3",
         "phpcompatibility/php-compatibility": "^9.3",
@@ -108,11 +108,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "App\\Tests\\": "tests/"
-        },
-        "classmap": [
-            "tests/_support"
-        ]
+            "Tests\\": "tests/"
+        }
     },
     "replace": {
         "paragonie/random_compat": "2.*",
