@@ -94,7 +94,7 @@ class RoomActivityStateSubscriber implements EventSubscriberInterface
             $event->setBlocked(true);
         }
 
-        // Deny, if community room has linked project rooms (this will also reset the account state)
+        // Deny, if community room has linked project rooms (this will also reset the room state)
         if ($room->isCommunityRoom()) {
             if ($this->roomManager->getLinkedProjectRooms($room)->getCount() > 0) {
                 $this->roomManager->resetInactivity($room, false, true, false);
