@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Tests\Api;
+namespace Tests\Api;
 
-use App\Tests\ApiTester;
+use Tests\Support\ApiTester;
 use Codeception\Util\HttpCode;
 
 class AccountsCest
