@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Tests\Api;
+namespace Tests\Api;
 
 use App\Entity\AuthSource;
 use App\Entity\AuthSourceShibboleth;
-use App\Tests\ApiTester;
+use Tests\Support\ApiTester;
 use Codeception\Util\HttpCode;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RouterInterface;
