@@ -1,10 +1,13 @@
 <?php
-namespace App\Tests\Helper;
+
+namespace Tests\Support\Helper;
 
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
 
-class Api extends \Codeception\Module
+use Codeception\Module;
+
+class Acceptance extends Module
 {
 
 }
