@@ -1,13 +1,10 @@
 <?php
-
-namespace App\Tests\Helper;
+namespace Tests\Support\Helper;
 
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
 
-use Codeception\Module;
-
-class Acceptance extends Module
+class Api extends \Codeception\Module
 {
 
 }
