@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Tests\Helper;
+namespace Tests\Support\Helper;
 
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
