@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Tests\Helper;
+namespace Tests\Support\Helper;
 
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
 
 use Codeception\Module;
 
-class Unit extends Module
+class Functional extends Module
 {
 
 }
